Moving switch logic to organization
