+++ app/models/book.rb
@@ -3,13 +3,6 @@ class Book < ActiveRecord::Base
 
   markdown_on :preface
 
-  before_create :setup_apartment_tenant!
-  after_destroy :teardown_apartment_tenant!
-
-  def switch!
-    Apartment::Tenant.switch! name
-  end
-
   def rebuild!(chapters)
     transaction do
       Chapter.all_except(chapters).delete_all
@@ -21,26 +14,5 @@ class Book < ActiveRecord::Base
     end
   end
 
-  def self.current
-    raise 'book not selected' if Apartment::Tenant.on? 'public'
-    find_by name: Apartment::Tenant.current
-  end
 
-  def self.central
-    find_by name: 'central'
-  end
-
-  def self.central?
-    current.name == 'central'
-  end
-
-  private
-
-  def teardown_apartment_tenant!
-    Apartment::Tenant.teardown name
-  end
-
-  def setup_apartment_tenant!
-    Apartment::Tenant.setup name
-  end
 end
