+++ app/views/chapters/index.html.erb
@@ -8,7 +8,7 @@
   </div>
 
   <div class="text-box">
-    <%= Book.current.preface_html %>
+    <%= Organization.current.book.preface_html %>
   </div>
 
   <div class="actions">
